Split image workspace query hydration
