+++ tests/image-workspace-split-contract.test.ts
@@ -7,6 +7,7 @@ const root = process.cwd();
 const imageDir = path.join(root, 'frontend/app/(core)/(workspace)/app/image');
 const workspacePath = path.join(imageDir, 'ImageWorkspace.tsx');
 const composerPersistenceHookPath = path.join(imageDir, '_hooks/useImageComposerPersistence.ts');
+const queryHydrationHookPath = path.join(imageDir, '_hooks/useImageWorkspaceQueryHydration.ts');
 
 const splitFiles = [
   '_components/ImageAuthGateModal.tsx',
@@ -15,6 +16,7 @@ const splitFiles = [
   '_hooks/useImageWorkspaceHistory.ts',
   '_hooks/useImageWorkspacePricing.ts',
   '_hooks/useImageWorkspaceDesktopLayout.ts',
+  '_hooks/useImageWorkspaceQueryHydration.ts',
   '_lib/image-workspace-character-references.ts',
   '_lib/image-workspace-copy.ts',
   '_lib/image-workspace-history.ts',
@@ -30,6 +32,7 @@ function readImageWorkspace() {
 test('image workspace foundations are split from the route orchestrator', () => {
   const source = readImageWorkspace();
   const composerPersistenceHookSource = readFileSync(composerPersistenceHookPath, 'utf8');
+  const queryHydrationHookSource = readFileSync(queryHydrationHookPath, 'utf8');
 
   for (const file of splitFiles) {
     statSync(path.join(imageDir, file));
@@ -41,6 +44,7 @@ test('image workspace foundations are split from the route orchestrator', () => 
   assert.match(source, /from '\.\/_hooks\/useImageWorkspaceHistory'/);
   assert.match(source, /from '\.\/_hooks\/useImageWorkspacePricing'/);
   assert.match(source, /from '\.\/_hooks\/useImageWorkspaceDesktopLayout'/);
+  assert.match(source, /from '\.\/_hooks\/useImageWorkspaceQueryHydration'/);
   assert.match(source, /from '\.\/_lib\/image-workspace-copy'/);
   assert.match(source, /from '\.\/_lib\/image-workspace-history'/);
   assert.match(source, /from '\.\/_lib\/image-workspace-utils'/);
@@ -58,13 +62,20 @@ test('image workspace foundations are split from the route orchestrator', () => 
   assert.doesNotMatch(source, /const remoteImageJobs =/, 'remote image history derivation belongs in useImageWorkspaceHistory');
   assert.doesNotMatch(source, /parsePersistedImageComposerState/, 'composer storage parsing belongs in useImageComposerPersistence');
   assert.doesNotMatch(source, /IMAGE_COMPOSER_STORAGE_KEY/, 'composer storage constants belong in useImageComposerPersistence');
+  assert.doesNotMatch(source, /const requestedJobId = useMemo/, 'query job hydration belongs in useImageWorkspaceQueryHydration');
+  assert.doesNotMatch(source, /const requestedEngineId = useMemo/, 'query engine hydration belongs in useImageWorkspaceQueryHydration');
+  assert.doesNotMatch(source, /Job settings snapshot missing/, 'snapshot application belongs in useImageWorkspaceQueryHydration');
 
   assert.match(composerPersistenceHookSource, /export function useImageComposerPersistence/);
   assert.match(composerPersistenceHookSource, /parsePersistedImageComposerState/);
   assert.match(composerPersistenceHookSource, /IMAGE_COMPOSER_STORAGE_DEBOUNCE_MS/);
   assert.match(composerPersistenceHookSource, /localStorage\.getItem\(IMAGE_COMPOSER_STORAGE_KEY\)/);
   assert.match(composerPersistenceHookSource, /localStorage\.setItem\(IMAGE_COMPOSER_STORAGE_KEY, serialized\)/);
+  assert.match(queryHydrationHookSource, /export function useImageWorkspaceQueryHydration/);
+  assert.match(queryHydrationHookSource, /const requestedJobId = useMemo/);
+  assert.match(queryHydrationHookSource, /const requestedEngineId = useMemo/);
+  assert.match(queryHydrationHookSource, /applyImageSettingsSnapshot/);
 
   const lineCount = source.split('\n').length;
-  assert.ok(lineCount <= 1300, `ImageWorkspace should stay below 1300 lines after composer persistence extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 1085, `ImageWorkspace should stay below 1085 lines after query hydration extraction, got ${lineCount}`);
 });
